+++ test/dart2_suggestors/pubspec_over_react_upgrader_test.dart
@@ -15,34 +15,9 @@
 import 'package:over_react_codemod/src/dart2_suggestors/pubspec_over_react_upgrader.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:test/test.dart';
-import 'package:yaml/yaml.dart';
 
 import '../shared_pubspec_tests.dart';
 import '../util.dart';
-
-/// Throws if [yaml] is an invalid pubspec, either due to:
-///
-/// - being unparseable
-/// - having incorrect structure (this check is not comprehensive)
-void validatePubspecYaml(String yaml) {
-  final yamlDoc = loadYamlDocument(yaml);
-
-  expect(yamlDoc.contents, isA<YamlMap>());
-  final extraTopLevelKeys =
-      (yamlDoc.contents as YamlMap).keys.toSet().difference(const {
-    'name',
-    'version',
-    'author',
-    'executables',
-    'description',
-    'dependencies',
-    'dev_dependencies',
-    'dependency_overrides',
-  });
-  expect(extraTopLevelKeys, isEmpty,
-      reason: 'unexpected top-level keys in pubspec.yaml;'
-          ' could the dependencies be missing indentation?');
-}
 
 main() {
   group('PubspecOverReactUpdater', () {
@@ -118,7 +93,15 @@ main() {
   });
 }
 
-String getExpectedOutput() {
+String getExpectedOutput({bool useMidVersionMin = false}) {
+  if (useMidVersionMin) {
+    return ''
+        'dependencies:\n'
+        '  over_react: ">=2.0.0 <3.0.0"\n'
+        '  test: 1.5.1\n'
+        '';
+  }
+
   return ''
       'dependencies:\n'
       '  over_react: ">=1.30.2 <3.0.0"\n'
