+++ test/shared_pubspec_tests.dart
@@ -13,38 +13,14 @@
 // limitations under the License.
 
 import 'package:meta/meta.dart';
+import 'package:over_react_codemod/src/util.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:test/test.dart';
-import 'package:yaml/yaml.dart';
 
 import 'util.dart';
 
-/// Throws if [yaml] is an invalid pubspec, either due to:
-///
-/// - being unparseable
-/// - having incorrect structure (this check is not comprehensive)
-void validatePubspecYaml(String yaml) {
-  final yamlDoc = loadYamlDocument(yaml);
-
-  expect(yamlDoc.contents, isA<YamlMap>());
-  final extraTopLevelKeys =
-      (yamlDoc.contents as YamlMap).keys.toSet().difference(const {
-    'name',
-    'version',
-    'author',
-    'executables',
-    'description',
-    'dependencies',
-    'dev_dependencies',
-    'dependency_overrides',
-  });
-  expect(extraTopLevelKeys, isEmpty,
-      reason: 'unexpected top-level keys in pubspec.yaml;'
-          ' could the dependencies be missing indentation?');
-}
-
 void sharedPubspecTest({
-  @required Function() getExpectedOutput,
+  @required Function({bool useMidVersionMin}) getExpectedOutput,
   @required SuggestorTester testSuggestor,
   @required String dependency,
   @required VersionRange startingRange,
@@ -142,7 +118,7 @@ void sharedPubspecTest({
             '  test: 1.5.1\n'
             '',
         expectedOutput: shouldUpdateMidRange
-            ? getExpectedOutput()
+            ? getExpectedOutput(useMidVersionMin: true)
             : ''
                 'dependencies:\n'
                 '  $dependency: $midVersionRange\n'
